+++ packages/e2e-test-app-fabric/test/TextInputComponentTest.test.ts
@@ -641,4 +641,10 @@ describe('TextInput Tests', () => {
     const dump = await dumpVisualTree('textinput-padding');
     expect(dump).toMatchSnapshot();
   });
+  test('TextInputs can have caretHidden', async () => {
+    const component = await app.findElementByTestID('textinput-carethidden');
+    await component.waitForDisplayed({timeout: 5000});
+    const dump = await dumpVisualTree('textinput-carethidden');
+    expect(dump).toMatchSnapshot();
+  });
 });